fix: address part comments
